fix(select): update floating ui logic
