+++ packages/react-magma-dom/src/components/Select/index.tsx
@@ -1,13 +1,6 @@
 import * as React from 'react';
 
 import {
-  AlignedPlacement,
-  autoUpdate,
-  flip,
-  useFloating,
-} from '@floating-ui/react-dom';
-import { ReferenceType } from '@floating-ui/react-dom/dist/floating-ui.react-dom';
-import {
   useMultipleSelection,
   UseMultipleSelectionProps,
   useSelect,
@@ -131,10 +124,6 @@ export interface SelectProps<T extends SelectOptions>
   /**
    * @internal
    */
-  floatingElementStyles?: React.CSSProperties;
-  /**
-   * @internal
-   */
   hasError?: boolean;
   /**
    * Reference to the trigger button element in the select
@@ -168,14 +157,6 @@ export interface SelectProps<T extends SelectOptions>
    * Event that will fire when a keypress is released while focused on the trigger button
    */
   onKeyUp?: (event: React.KeyboardEvent) => void;
-  /**
-   * @internal
-   */
-  setFloating?: (node: ReferenceType) => void;
-  /**
-   * @internal
-   */
-  setReference?: (node: ReferenceType) => void;
 }
 
 export interface MultiSelectProps<T extends SelectOptions>
@@ -261,35 +242,23 @@ export function Select<T>(props: XORSelectProps<T>) {
 
   const isInverse = useIsInverse(props.isInverse);
 
-  const { floatingStyles, refs } = useFloating({
-    middleware: [flip()],
-    placement: 'bottom-start' as AlignedPlacement,
-    whileElementsMounted: autoUpdate,
-  });
-
-  const floatingElementStyles = { ...floatingStyles, width: '100%' };
-
   return (
     <div style={containerStyle} data-testid={testId}>
       {isMulti && instanceOfMultiSelect<T>(props) ? (
         <MultiSelect
           ariaDescribedBy={descriptionId}
-          floatingElementStyles={floatingElementStyles}
           hasError={hasError}
           id={id}
           isInverse={isInverse}
           itemToString={itemToString}
           labelPosition={labelPosition || LabelPosition.top}
           labelWidth={labelWidth}
-          setFloating={refs.setFloating}
-          setReference={refs.setReference}
           {...(props as MultiSelectProps<T>)}
         />
       ) : (
         <InternalSelect
           ariaDescribedBy={descriptionId}
           errorMessage={errorMessage}
-          floatingElementStyles={floatingElementStyles}
           hasError={hasError}
           helperMessage={helperMessage}
           id={id}
@@ -298,8 +267,6 @@ export function Select<T>(props: XORSelectProps<T>) {
           labelPosition={labelPosition || LabelPosition.top}
           labelWidth={labelWidth}
           messageStyle={messageStyle}
-          setFloating={refs.setFloating}
-          setReference={refs.setReference}
           {...(props as SelectProps<T>)}
         />
       )}
