+++ packages/react-magma-dom/src/hooks/useMagmaFloating.ts
@@ -1,0 +1,14 @@
+import {
+  useFloating,
+  flip,
+  autoUpdate,
+  AlignedPlacement,
+} from '@floating-ui/react-dom';
+
+export function useMagmaFloating(placement: AlignedPlacement = 'bottom-start') {
+  return useFloating({
+    middleware: [flip()],
+    placement,
+    whileElementsMounted: autoUpdate,
+  });
+}
